Fix 32-bit build issue
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1701 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ test/h5file/h5varlen.cpp
@@ -135,7 +135,7 @@ void test_var_length_data_common( const char* filename, MBInterface& mb1, bool o
     CHECK( NULL != dataptr );
       // check size 
     CHECK( size > 0 );
-    CHECK_EQUAL( 0ul, size % sizeof(int) );
+    CHECK_EQUAL( 0ul, (unsigned long)(size % sizeof(int)) );
     size /= sizeof(int);
     CHECK_EQUAL( size, dataptr[0] );
       // check other values
